Changed auxiliaryKwh and batterycapacity fields to double

+++ ElectroWayFINAL/src/main/java/com/example/electrowayfinal/models/Car.java
@@ -36,8 +36,8 @@ public class Car {
     private long year;
 
     @Basic
-    @Column(name = "battery_capacity", nullable = false, columnDefinition = "bigint")
-    private long batteryCapacity;
+    @Column(name = "battery_capacity", nullable = false, columnDefinition = "double")
+    private double batteryCapacity;
 
     @Basic
     @Column(name = "charging_capacity", nullable = false, columnDefinition = "double")
@@ -52,8 +52,8 @@ public class Car {
     private long vehicleMaxSpeed;
 
     @Basic
-    @Column(name = "auxiliary_kwh", nullable = false, columnDefinition = "bigint")
-    private long auxiliaryKwh;
+    @Column(name = "auxiliary_kwh", nullable = false, columnDefinition = "double")
+    private double auxiliaryKwh;
 
     @ManyToOne
     @JoinColumn(name = "owner_id", referencedColumnName = "id", columnDefinition = "bigint")
